split arguments -> arguments + channel
